# balances/src/main/resources/openapi/api.yml
openapi: 3.0.0

info:
  title: Balance Microservice API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api
    description: Local Server

paths:
  /balance/v1/balances:
    get:
      summary: Get balance details
      operationId: balances
      tags:
        - balances
      responses:
        '200':
          description: Balances information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalancesResponse'

  /accounts/{AccountId}/balances:
    get:
      summary: Get balance details by Id
      operationId: balancesById
      parameters:
        - name: AccountId
          in: path
          description: Account ID
          required: true
          schema:
            type: integer
            format: int64
      tags:
        - balancesById
      responses:
        '200':
          description: Balance detail by Id information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceByIdResponse'

  /balance/v1/health-check:
    get:
      summary: Balance Health Check
      operationId: health-check
      responses:
        '200':
          description: Checking Balances MS Health
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Not Found
          content: { }

components:
  schemas:
    BalanceByIdResponse:
      type: object
      properties:
        Data:
          type: object
          properties:
            Balance:
              type: object
              properties:
                AccountId:
                  type: string
                Amount:
                  type: object
                  properties:
                    Amount:
                      type: string
                    Currency:
                      type: string
                CreditDebitIndicator:
                  type: string
                Type:
                  type: string
                DateTime:
                  type: string
                  format: date-time
                CreditLine:
                  type: array
                  items:
                    type: object
                    properties:
                      Included:
                        type: boolean
                      Amount:
                        type: object
                        properties:
                          Amount:
                            type: string
                          Currency:
                            type: string
                      Type:
                        type: string
        Links:
          type: object
          properties:
            Self:
              type: string
        Meta:
          type: object
          properties:
            TotalPages:
              type: integer
              format: int32

    BalancesResponse:
      type: object
      properties:
        Data:
          type: object
          properties:
            Balance:
              type: array
              items:
                type: object
                properties:
                  AccountId:
                    type: string
                  Amount:
                    type: object
                    properties:
                      Amount:
                        type: string
                      Currency:
                        type: string
                  CreditDebitIndicator:
                    type: string
                  Type:
                    type: string
                  DateTime:
                    type: string
                    format: date-time
                  CreditLine:
                    type: array
                    items:
                      type: object
                      properties:
                        Included:
                          type: boolean
                        Amount:
                          type: object
                          properties:
                            Amount:
                              type: string
                            Currency:
                              type: string
                        Type:
                          type: string
        Links:
          type: object
          properties:
            Self:
              type: string
        Meta:
          type: object
          properties:
            TotalPages:
              type: integer
              format: int32